+++ tests/jax/test_likelihood.py
@@ -11,11 +11,9 @@ import numpy as np
 import numpy.testing as npt
 
 from tests._shared.noisemodel_helpers import build_fd_pair, dense_kernel_2ch
-from tests._shared.waveforms_helpers import \
-    build_harmonic_projected_frequency_waveform
+from tests._shared.waveforms_helpers import build_harmonic_projected_frequency_waveform
 from typed_lisa_toolkit.consumers.likelihood import FDWhittleLikelihood
-from typed_lisa_toolkit.consumers.noisemodel import (FDNoiseModel,
-                                                     SpectralDensity)
+from typed_lisa_toolkit.consumers.noisemodel import FDNoiseModel, SpectralDensity
 from typed_lisa_toolkit.containers.data import FSData
 from typed_lisa_toolkit.containers.representations import FrequencySeries
 from typed_lisa_toolkit.containers.waveforms import sum_harmonics